Seprated Play and Test logic and functions to util functions
