+++ components/types.d.ts
@@ -1,0 +1,6 @@
+interface Maps {
+    _id: Id<"maps">;
+    _creationTime: number;
+    level: number;
+    grid: string[][];
+}
